GafferImageUI : Moved ChannelPlugValueWidget from ShuffleUI into being something reusable
